Add a new Visualization

+++ PaperDetection/main.py
@@ -1,10 +1,12 @@
 import sys
 from capture import Capture
 
+cp = None
 try:
     cp = Capture()
     cp.run()
-except:
-    cp.terminate()
+except Exception as e:
+    if cp != None: cp.terminate()
+    print e
     print "Bye"
     sys.exit()
